Fix: bugs de compilation avec Opus 4 Partie 2 espoir 2

+++ src/main/java/com/regulation/contentieux/model/Encaissement.java
@@ -109,7 +109,7 @@ public class Encaissement {
     /**
      * Alias pour getMontantEncaisse() - compatibilité RapportService
      */
-    public BigDecimal getMontant() {
+    public long getMontant() {
         return getMontantEncaisse();
     }
